third_party/blink/web_tests/platform/mac-mac10.11/virtual/outofblink-cors/external/wpt/service-workers/service-worker/clients-matchall-client-types.https-expected.txt

This is a testharness.js-based test.
FAIL Verify matchAll() with window client type assert_array_equals: property 1, expected true but got false
FAIL Verify matchAll() with {window, sharedworker, worker} client types promise_test: Unhandled rejection with value: object "Error: wait_for_state must be passed a ServiceWorker"
Harness: the test ran to completion.
